Service price changed to decimal. Code clean up from usings.

+++ SmartGarage/SmartGarage.Service/DTOs/GetDTOs/GetServiceDTO.cs
@@ -13,6 +13,6 @@
 
         public string Name { get; }
 
-        public double Price { get; }
+        public decimal Price { get; }
     }
 }
